slurm_completion - fix sacctmgr handling
This patch fixes and changes a few things:

(1): A completion bug that was boarderline a feature.

Example:
$ sacctmgr modify user <user>

This is valid for Slurm to take, but it was a bug within the logic flow.

(2): Better parameter gating with 'where' and 'set' conditions.

Parameters are only suggested for the last found condition.

(3) Better match the CRUD documentation.

- sacctmgr delete/remove <ENTITY> where <SPECS>
- sacctmgr list/show <ENTITY> [<SPECS>]
- sacctmgr add/create <ENTITY> <SPECS>
- sacctmgr modify/update <ENTITY> where <SPECS> set <SPECS>

list/show and add/create relied on 'where' condition despite not
requiring it.

Bug 17906