Handle extra constraints satisfied on some but not enough nodes
This is a continuation of commit 9f4ecc22d4.

Do not hold batch jobs whose extra constraints are satisfied on at least one
node, but not enough nodes to fulfill the job request. Set these jobs to the
state reason FAIL_CONSTRAINTS.

This fix is a very simplistic check if the job cannot be scheduled and the
job requested extra constraints. We do not actually know if the job cannot
run due to the extra constraints or for some other reason.

Bug 18064