Fix race on reconfig variable.
A ill-timed 'scontrol reconfigure' or SIGHUP received before the
"reconfig = false" setting would lead to the slurmctld shutting down
rather than attempting to reconfigure.

Avoid this by using a new temporary variable for the setup side of
the controller operations, distinct from the shutdown and reconfig
handling portion of the code. We can then safely clear the reconfig
variable when the signal handler and RPC managing threads aren't running.

Bug 18214.

Signed-off-by: Brian Christiansen <brian@schedmd.com>