Fix crash that can occur when launching via pmi2 with route/topology
slurm_conf.max_node_cnt must be initialized to NO_VAL (and not 0) in
slurmstepd to avoid a crash when calling grow_node_record_table_ptr().
057ba0413e assumed node_record_count is non-zero when entering that
function but it can be when slurmstepd is called during pmi2 launching.

Set to NO_VAL to be consistent with the way max_node_cnt is used
elsewhere and to avoid the crash.

Bug 18002