Ensure install location of Slurmdb.pm is valid.
slurm.spec and seff depend on libslurmdb install in _perldir as defined in:
contribs/seff/Makefile.am and slurm.spec. In some sites, MakeMaker uses a
different default value for INSTALLSITE{LIB,ARCH}, and libslurmdb will not be
found by seff (and/or slurm.spec). With the value INSTALLSITE{LIB,ARCH} set in
the same manner as in seff and slurm.spec, libslurmdb will surely be found.

Bug 17817