Use the loop to look up all SlurmctldHost entries.
Previously a lookup failure on the primary controller would lead to the
backups being skipped. The working assumption was that this would only
happen on a DNS failure, but for cloud environments where the DNS entry
is tied to the host's state this proves problematic and prevents any
attempt to connect to the backup.

Bug 18280