data_parser/v0.0.41 - change ASSOC_ID parser to be complex
When parsing ASSOC_ID -> ASSOC_SHORT (previously psp type parser), we
not only want the id number, but we also want the cluster name. To do
this, the ASSOC_ID parser needs to be complex.

This is similar to what was done for ASSOC_ID (now JOB_ASSOC_ID) in
commit 8b6871946a.

Continuation of 9fecbeb4ca merging to v0.0.41.

Bug 18394