Avoid refreshing the hwloc xml file on slurmd reconfigure.
This fixes an issue seen when using CoreSpecCount on a node with Intel
E-cores where the reconfigured child would only see the CoreSpec CPUs,
and then fail to register to slurmctld. When the parent slurmd is running
under a restricted cpuset, hwloc_cpukinds_get_nr() does not return what
we expect - and there is not yet a flag to request it operate on the
whole node and not the restricted cpuset. So the end result of
_remove_ecores() was the cpuset as limited by CoreSpecCount.

Adding this new call - instead of adding this as an argument to
xcpuinfo_init(), sidesteps a quirk of the current xcpuinfo interface.
The init is not actually getting called first. Instead,
xcpuinfo_hwloc_topo_get() is usually the first to get into here.

Bug 18411.

Signed-off-by: Danny Auble <da@schedmd.com>