Only show "restart slurmctld" error message in slurmctld
If the slurmscriptd has a broken pipe, it means that its parent slurmctld
closed the connection, most likely because it stopped. Without this commit,
this error message showed every time that scontrol reconfigure happened
successfully, because on reconfigure the old slurmctld does not wait for
the old slurmscriptd to finish. This error only needs to be displayed if
it happens in slurmctld, in which case its slurmscriptd died for some reason
and the slurmctld needs to be restarted to recover.

Bug 14010