Automatically convert from cons_res to cons_tres when unpacking
select/cons_res was removed in 23.11. This forces a change to
select/cons_tres.

However, when a 23.11 slurmctld attempts to unpack state that had cons_res,
it cannot find the plugin because cons_res does not exist, resulting in an
unpack failure and a slurmctld fatal error. Changing the plugin id from the
old cons_res to cons_tres allows the unpacking to continue as if cons_tres
was used. This is done in the SLURM_MIN_PROTOCOL_VERSION block because it
only applies to state before 23.11.

Bug 18448