Remove faulty xassert in _calc_arbitrary_tpn and replace it with error
If there are duplicates in the job_ptr->details->req_nodes then the
xassert would fail, but the hostlist is meant to contain duplicates in
this case.

The number of unique hosts in the hostlist must be equal to num_nodes.
Return an error if that is not the case. This is needed because if the
min_nodes is larger then the number of nodes in the hostlist then the
ntasks_per_node array will have to many entries and will be ordered
incorrectly. But if the min_nodes is less then the number of nodes in
the nodelist then the slurmctld will crash when calculating the
tasks_per_node array.

Bug 18422