Abort slurmctld takeover if clock skew is detected.
If the backup controller is sufficiently ahead in time from the primary,
it would be possible to start running despite the primary still operating.

This change will prevent the takeover if the heartbeat file is being
updated, even if the backup can't ping the primary. Taking over while
the primary is running causes problems.

Bug 18546