jobacct_gather/common - aggregate completed processes only once
Instead of keeping all processes from the job in prec_list and aggregate
everything in every iteration we can permanently aggregate the details
about completed processes. This way going over the list may be much
faster and we reduce the risk of data being overridden - if the PID is
reused.

Bug 18568