Avoid generating a query larger than max_allowed_packet
With large association tables, it is possible to generate a very large query
to set lineage when updating the <clustername>_assoc_table. This is
particularly problematic with older mysql servers with a default
max_allowed_packet of 1 MiB. Split the update query into multiple queries if
needed to ensure that each query is smaller than max_allowed_packet.

Bug 18577