Don't allow deleting nodes defined in slurm.conf
Only allow deleting dynamic norm nodes. This prevents holes in the
node table in the range of the nodes of read in from the slurm.conf and
later being filled in by dynamic nodes which can create sorting issues
when the nodes are read back in from the slurm.conf on start -- see
previous commit.

Bug 17858

Signed-off-by: Scott Hilton <scott@schedmd.com>