burst_buffer/lua - xassert on issues with unpacking job_info
The previous commit fixed an issue with unpacking job info which is
passed to burst_buffer.lua. This bug was missed because the only way to
see the bug was to attempt using it in burst_buffer.lua. That is unlikely to
be tested. This commit adds an assert that should help to catch unpacking
issues before passing job info to burst_buffer.lua.

Bug 18603