cgroup/v2 - do not account for Virtual Memory in jobacct_gather/cgroup
The returned value was an unreliable approximation, and kernel cgroup
interfaces do not provide the virtual set size metric anyway. Getting
this value from /proc/<pid>/stat and mixing it with the actual cgroup
metrics didn't seem a good solution either.

Return 0 for VmSize and add a limitation section in the docs.

Bug 17594