Change error() messages in xgetaddrinfo() to error_in_daemon().
Failure to resolve compute node addresses or slurmctld hosts is not
likely to be something that the user can address.

This is readily obvious in an HA setup where the DNS entry for a
SlurmctldHost has been withdrawn. While the client can still
communicate successfully, the error() messages are distracting.