Fix checking if a script has been killed after track_script_flush
track_script_flush sets flush_called to true. In slurmctld or slurmscriptd,
this is called on a reconfigure or shutdown. If reconfigure fails, then
operation continues. However, any subsequent scripts where
track_script_killed was used (which is most of the scripts that run in
slurmctld or slurmscriptd) would be killed immediately. For example:

* run_command is used to run a script
* run_command_poll_child uses track_script_killed to know whether to break
  out early
* After the loop ends, the script is killed immediately with SIGTERM and
  SIGKILL because send_terminate defaults to true

Rather than use flush_called to know which list to look for the track script
record, look for the record in both lists. This avoids the problem with
flush_called altogether.

Bug 18489