Fix building the correct alias_list for jobs on startup
Currently, the alias_list isn't state saved and needs to be rebuilt on
startup the in the same way it was initially created at allocation time.
For example, if the job is still configuring and waiting for nodes to
boot, then the alias_list should have "TBD" so that when the nodes are
all booted the alias_list will be built with the correct node addresses.

Bug 18721

Signed-off-by: Skyler Malinowski <skyler@schedmd.com>