Add type_name and type_id to gres_step_state_[un]pack().
Previously the type_name and type_id of the step's gres_step_state_t were
being lost when the slurmctld either restarted or reconfigured. This could
cause a "dealloc count underflow" error in _step_dealloc() when the step
finished since it couldn't search the step's gres list for the correct gres
type.

This solves the issue by packing the gres_step_state_t type_name and
unpacking it. In the unpack function the type_id is built using the
type_name.

Bug 18836