Always allocate at least one gpu of each requested type per node.
When using --gpus=typeA:#,typeB:#,... allocate at least one of each gpu
type on each node of the allocation. Before depending on your gres.conf
an allocation would either only have the gpus requested or enforce having
one of each type on the node. When it did not enforce it the following
error would be in the slurmctld logs:
"error: _set_job_bits1: job # failed to find any available GRES on node #"
This would occur because pick_gres was set to 0 and all gpus on the
allocated socket and gpus not bound to a socket would be skipped. So if
there weren't any gpus of that type on non-allocated sockets then none of
that gpu type would be allocated on that node.

This fixes the issue by setting pick_gres to a minimum of 1. That way the
behavior is consistent regardless of how gres.conf specifies gpus are bound
to sockets.

Bug 18655