Do not fatal() when using scontrol to create a node with gres autodetect
When gres.conf has a global autodetect setting and scontrol create node is
used, slurmctld would fatal() because autodetect is not allowed for dynamic
nodes. This means an errant scontrol create node could cause a running
slurmctld to fatal(). To avoid accidentally downing the slurmctld, log and
return an error instead of fatal().

Bug 15692

Signed-off-by: Brian Christiansen <brian@schedmd.com>