Always map dynamic future nodes based off of the hostname
By default, the slurmd sets the slurmd's nodename to the hostname. If
-N<name> is specified, the initial check to see if the node exists is
done by the given nodename. But if the slurmd is restarted with the
-N<name> then it won't find the existing mapping and will map to
different node.

To avoid getting into this situation, always map to the hostname --
which is what it was doing by default when slurmd wasn't started with
-N<name>. This enforces it on the controller side.

Bug 18557

Signed-off-by: Ben Glines <ben.glines@schedmd.com>