Don't allow dynamic future nodes to specify a node name
Because it doesn't do anything and causes problems.

Specifying a node name doesn't dictate the name of the node since the
controller will map a future node's node name to the slurmd's hostname.

By default, when not specifying a name with -N, the node registers with
the controller using it's hostname as the nodename. The controller maps
an available future node to the slurmd registration. The future node's
hostname is set to the slurmd's hostname.

Subsequent periodic registrations from the slurmd register with the
mapped nodename so the node is found by the mapped future nodename.

On restart, the slurmd registers with the hostname and the controller
finds the node by the hostname.

If a dynamic future node initially starts with a specified node name
(-N<name>), the controller won't find a node by the given name and maps
a future node to the slurmd's hostname. But the specified node name is
ignored.

On a reconfig/restart with -N<name>, the node will register with the
given node name but the controller won't find a node with the given node
name (because the slurm.conf node is mapped to hostname) so it will map the
registration to another future node.

Bug 18557

Signed-off-by: Ben Glines <ben.glines@schedmd.com>