Fix controller not validating periodic dynamic future registrations
The controller periodically asks slurmds to send a new registration
(every 20 pings). When the slurmd responds to these requests, it asks
for the controller to send a response message with extra information
(e.g. TRES). Dynamic future nodes use the response message to send the
mapped future node name to the slurmd. For dynamic future node
registrations asking for the response, the controller never validates
the node registration (not needed on the first slurmd handshake to get
the mapped node name); so it wasn't validating the node registration for
periodic registrations after the mapping had taken place. So things like
slurmd load average would never be updated.

The solution is to validate the registration if the controller finds an
existing node record for the registration -- meaning we've already
mapped the node.

Bug 18557

Signed-off-by: Ben Glines <ben.glines@schedmd.com>