Fix tracking of exclusive multi-node jobs total GRES.
As mentioned in the comment from commit 30468d0c21, once the nodes are
selected for the job, the gres_job_state total_gres is interpreted as
the count of this GRES for the job (entirely aggregating the GRES from
all selected nodes).

This partially reverts what is now considered a regression introduced in
23.02.4 commit d62c31da95, since such commit prevented the aggregation from
happening.

When _job_select_whole_node_internal() is called the selection of nodes
has already happened, and if total_gres is not aggregated encompassing
all selected nodes the job could potentially bypass limits configured
for such GRES (as the count is lower than what it actually should).

Bug 18812