error_in_daemon() if UsePSS or NoShare are set with an incompatible plugin
jobacct_gather/cgroup or /none are not compatible with UsePss or NoShare
as the memory calculations are different/non-existent. In this case the
common_jag.c code is not as common as expected and is specific to the
linux plugin and procfs metrics. jobacct_gather/linux is the only one
at the moment that works with such parameters.

Bug 10587