Move unsafe call closeall() out of child in _get_user_env
_get_user_env called env_array_user_default which in turn forked a new
process to get the user environment. We forked from slurmd and already being
multithreaded. After the fork/clone and before the exec of 'su', there were
several calls which were unsafe and e.g. used malloc, leading to deadlocks.

This commit moves the closeall() which in turn called opendir() out of the
child.

Ideally we should just exec after fork, but as the code is set now we need
to close all the open file descriptors if we don't want them inherited by
the exec. These are possibly unsafe operations too. In this patch we
modified the code to just use the syscall close() (at expenses of
performance), and skipping opendir() and friends to try to mitigate the
risk.

Bug 18742
Bug 18838

Signed-off-by: Ben Glines <ben.glines@schedmd.com>