Move unsafe call env_array_overwrite out of child in _get_user_env
_get_user_env called env_array_user_default which in turn forked a new
process to get the user environment. We forked from slurmd when it was
multithreaded. After the fork/clone and before the exec of 'su', there were
several calls which were unsafe and e.g. used malloc, leading to deadlocks.

This commit fixes the calls to:
env_array_create()
env_array_overwrite()

The deadlock happened because malloc sets internal mutexes and before
the fork there were other threads holding these mutexes. In the child the
threads were gone but the mutexes were not released.

Forking from a multithreaded app is dangerous and malloc and other unsafe
functions *cannot be called*.

Bug 18742
Bug 18838

Signed-off-by: Ben Glines <ben.glines@schedmd.com>