Revert "Check for passed deadline on jobs being held after a finished resv"
This reverts commit 01ea23384e68250e7d688e26dd41dafc5d5bc1d1.

This commit is now redundant due to previous commit in combination with
commit 0145a50917 and can be reverted.

After the aforementioned two commits, the slurmctld periodically checks
in background for jobs --deadline handler, thus all pending jobs (held or
not) are checked against --deadline. This makes the particular care
introduced in 01ea23384e unneeded/redundant.

Bug 18894