Set num_tasks and min_nodes when submitting an arbitrary job with restd
Since slurmrestd can submit an arbitrary job that leaves min_nodes and or
num_tasks as NO_VAL, when slurmctld is validating the request in
validate_job_create_req() both values need to be set apropriatly. The
min_nodes value must be the unique count from the req_nodes list and
num_tasks must be the full count of req_nodes list.

Bug 18456