If job is too old, set version to SLURM_PROTOCOL_VERSION
We want to maintain job_ptr->start_protocol_ver in case we need to send
messages to an older client (e.g. x11 forwarding to an older srun client
that wouldn't be able to receive RPCs if packed with the slurmctld current
version).

If the job_ptr->start_protocol_ver is too old, we won't be able to pack any
messages to send to the older srun client anyways, so is reset to the
current verison.

Bug 18973