Add ntasks_opt_set parameter for explicitly set ntasks
For quite some time, "slurm_opt_t" structs have contained a "ntasks_set"
field which indicated when the user has explicitly requested a specific
amount of tasks using the --ntasks option in the cli. However, this field
ended up being used even for indicating when the number of tasks had been
implicitly calculated in different parts of the code.

Add a new "ntasks_opt_set" field to "slurm_opt_t" which inherits the
original purpose of "ntasks_set", preserving the actual one in order to
keep functionalities that have been built upon it. It provides a way to
actually identify when a user explicitly requested -n.

Bug 18217