srun - Removed "can't honor" warning on job allocations
Previously, when the combination of -N/-n/--ntasks-per-node wasn't
coherent, we just warned the user and ignored --ntasks-per-node.
With this, we remove this behavior for job allocations directly done
with srun and we just reject the job if the parameters aren't consistent.

This behavior however is preserved for step allocations, since
ntasks-per-node is not used there as per documentation. This change
doesn't affect the "can't honor" warning for allocations with -mpack.

This commit also narrows the conditions where the warning and reset
should actually happen.

Bug 18217