run_command - Add plumbing for the script launcher
If a caller specified a binary with run_command_init(), then this binary
will exec'd instead of the script, with argv[1] being a special string to
indicate that the binary was launched in script mode. The script path and
argv are passed in argv. The launcher is executed with the script's
environment if provided to run_command.

Ticket 16126