Down the node if helper script returns non-zero return code
We first send a registration to clear the powering_up flag. The node
will go into a INVAL state because the current features won't be a super
set of the current active features on the controller. If we don't first
send the registration then the node will hit ResumeTimeout and the node
will be marked down and any previous reason will be lost. We then mark
the node down with a reason.

Bug 12288