Fix issues with tasks when a job was shrinked more than once.
Shrink num_tasks if there are not enough cpus in the job allocation.

If a running job is shrunk with NumNodes, the nodes and cpus will shrink,
but the number of tasks will not shrink. The next attempted shrink on
the job will be rejected because there are more tasks than cpus.

Bug 18715