burst_buffer/lua - Remove stage throttling
burst_buffer/lua was introduced in 21.08. Around the same time, commit
18b2f4fff3 was introduced in 21.08 which placed a cap of 4096 open files on
slurmctld (and thus also slurmscriptd). That commit was later reverted, but
the limit in burst_buffer/lua was not ever reverted. The removal of the 4096
open file limit means that this throttling is likely not needed. With the
addition of slurm_bb_test_data_{in|out}, it is not necessary to keep the
data_{in|out} hook running for a long time, so the number of open files can
be reduced.

Therefore, this commit removes the stage throttling in burst_buffer/lua.

Bug 17924