TRES requested by tres-per-task using the old format will not fail
If a custom gres or license was defined with its name ending in
'gres' or 'license' respectively, the request would fail as
xstrsubstituteall replaced all instances of 'gres:' or license:'
with 'gres/' or 'license/'. The new slurm_format_tres_string will
correctly change the old request format to the new request format.

Bug 18532