burst_buffer plugins - Remove __func__ from many logs
When a plugin logs at info level or more verbose, the function name is
automatically prepended to the message. So, logging the function name was
redundant. This commit removes it in most places in burst_buffer_common.c.
There are a few places that are intentionally left unchanged, since those
are only used by the datawarp plugin which is not being updated anymore.