Properly wait for a scope to be initialized in slurmd init and cgroup/v2
In cgroup/v2 slurmd will startup a systemd scope if it is not already
started. We need to properly wait for the cgroup directories to be created
plus the pid of our child to be moved there before continuing any setup
operation in the cgroup interfaces, or we might race with systemd.

From when we send the call to dbus to create the scope untill everything
is ready, a lot of time can pass. Systemd will first create its internal
structures, then mkdir the cgroup directories, and finally move the pid
specified in the dbus call to the cgroup.procs of the root. We need to
avoid any race in this asynchronous process, so we need to wait.

Bug 19157