Modify _set_max_gres to be aware of rem_nodes and rem_sockets.
This sets the max_gres for gres_per_job to the max this node could have. It
saves at least one gpu for each remaining nodes. If total_gres is already
more then or equal to gres_per_job allow the min gres on this node with
respect to gres_per_node, gres_per_sock, and gres_per_tasks.

Bug 14227