data_parser/v0.0.41 - Add parser for job/submit distribution
Originally the PARSER_ARRAY(JOB_DESC_MSG)[] used the TASK_DISTRIBUTION
parser which had its parsing function disabled. This adds the new
JOB_DESC_MSG_TASK_DISTRIBUTION which is needed when parsing a task
distribution because it also has to set the plane size value for the job
description if it is specifed.

Bug 18456