Fix sattach.
Credential refactoring work ahead of 23.11 inadvertently broke sattach for
non-SlurmUser/root accounts. The change made slurmd unwilling to unpack a
launch credential that did not pass verification. Unfortunately, sattach was
using a "fake" credential, and relied on the credential decoding (but not
passing verification, which is now baked-in to the decode) in order to
obtain the credential signature. The signature of this credential served as
the io_key, and no other fields were checked or referenced.

Rather than reverting portions of the cred cleanup, explicitly change the
REQUEST_REATTACH_TASKS to send the desired io_key. This is an RPC-breaking
change which is generally banned on the stable releases. - but the fact that
sattach is completely broken as-is provides some degree of justification
for this approach.

Bug 18744.

Signed-off-by: Tyler Connel <tyler@schedmd.com>