Docs - Update section referring to pam_access
We described one direction for stacking it, but don't fully describe
that you can stack them both ways and what the implications of that are.
Provide examples with behavior changes depending on how the modules are
stacked.

pam_access is also capable of specifying specific users, so don't imply
that it doesn't.

Bug 19354

Signed-off-by: Stephen Kendall <stephen@schedmd.com>