Avoid limiting sockets per node to one when using gres enforce-binding.
Originally --cpus-per-gpu was not taken into account when deciding if the
number of sockets per node should be limited to one. This would case jobs
like 'srun --gpus=2 --cpus-per-gpu=1 --gres-flags=enforce-binding hostname'
to be rejected if there was only one gpu per socket.

This fixes the issue by getting the minimum gres cpus needed by the job.
The resulting cpu count is not necessarily accurate for the final cpu per
node count since gres per node might be estimated from gres_per_job and
min_nodes. But it will tell if the required cpu count is more than one. If
it is more than one, then it does not limit the sockets per node to 1.

Bug 18425