Move the calcuation of ntasks from srun to slurmctld when -c without -n
Before when -c was used without -n srun would calculate the ntasks to be
the maximum number of tasks that could run in the full allocation. This
calculation would disregard -N and --threads-per-core/--hint=nomultithread.

Now the ntasks calculation is deferred by setting it as NO_VAL. Since
ntasks is used to set cpu_count that is also set to NO_VAL. These values
get passed to the slurmctld in the step_req by launch_g_create_job_step().
Slurmctld sets num_tasks and cpu_count in _pick_step_nodes() after the
avail_nodes have been set. num_tasks is set to the maximum number of tasks
that the required number of nodes can run while consiering threads per
core. This does mean it doesn't account for some cpus already be in use by
another step, which might cause the step to hang until the cpus are freed.
However if a range is given for -N, then if the min_nodes is satificed by
the count of idle nodes those nodes will be used to calculate num_tasks.

Bug 18550