Fix incorrect exclusive allocation for certain multi-part jobs.
This could happen if a multi-part job doesn't specify --exclusive and is
first evaluated on a partition configured with OverSubscribe=EXCLUSIVE but
ends up being allocated in a partition configured without EXCLUSIVE that is
evaluated afterwards.

select_nodes() eventually calls _resolve_shared_status(), which alters the
job_ptr->details->{whole_node,share_res} of the job. Unless the call
returns SUCCESS and test_only is false (meaning the job is allocated
resources in the evaluated partition) the fix is to restore the original
values otherwise. Since select_nodes() is the first common ancestor
function to _resolve_shared_status(), this is the function considered
more appropriate to add the save/restore logic.

Bug 19196