Setting DebugFlags=GLOB_SILENCE flag no longer exposes old behavior
The GLOB_SILENCE flag added in commit 651a2bc6dd was intended to silence
a message that would display when a glob symbol was detected when
parsing include directives. Setting the flag caused it to instead try to
read the glob symbol as a file due to a poorly constructed if statement.

Bug 18762, 18347