Testsuite - Rename default node to node0 instead of node1
This way if require_nodes() is not called tests will use node0 only, and
when used, the first node will be node1 (instead of node2).

This is a continuation of previous commit just to avoid confusions, and
needed some tests to be adapted and a change in the containers too.

Bug 18464