Reduced restricted gpu cores based on gres layout during node selection.
As nodes are selected for the job allocation mock laying out tasks and
gres. If removing unused gpus and there associated restricted cores on a
node reduces the cpu count below the minimum required by the job don't
select the node or reject the job if the node was required.

This does have the side effect of making _eval_nodes_consec skip some
nodes in a consecutive group if it is found the node doesn't have enough
resources due to gres layout.

Bug 14227