Remove unused function parse_int_to_array
parse_int_to_array() and hostlist_parse_int_to_array() were both added in
commit d82585cd90d. However, parse_int_to_array() was only added as a
function declaration without a definition and was unused, while
hostlist_parse_int_to_array() was added and used. It was most likely an
accident that a declaration for parse_int_to_array() was added but not used.