Testsuite - Fix test4.14 for multiple slurmd
e1b83ada8b changed it so that partitions would fail to create with
invalid nodelists. test4.14 creates partitions with a hostname of an
idle node. In multiple slurmd, the hostname isn't a valid nodename.
Before the commit, the partition would still be created since it didn't
reject the creation of the partition - i.e. the partition would be
created with a blank nodelist. Now that valid nodelists are enforced the
partition wouldn't be created. The solution is to find nodes by nodename
and not hostname.

Bug 18034