rest API - Initialize flags to zero for DELETE jobs
Continuation of 240f205b94 by removing all initialized flags at all.

When ScronParameters=explicit_cancel is configured, the --cron option for
scancel has to be used to signal cron jobs. scancel sets the KILL_NO_CRON
flag if the --cron option is not requested.

The data parser plugins do not unset bits in a flag variable, they only set
the bits in a flag variable that were also set in the request. Thus commit
240f205b94 made it impossible for DELETE jobs to signal cron jobs with
explicit_cancel configured since it was impossible to unset the KILL_NO_CRON
flag. This is fixed by initializing the flags variable to zero.

Unfortunately, this means that the default flags value becomes the
equivalent of the scancel --cron option, which defeats the point of
explicit_cancel requiring an option to enable signalling cron jobs, but
that is better than completely disabling the ability the signal cron jobs
from the rest API.

Bug 18464

Signed-off-by: Nathan Rini <nate@schedmd.com>