Add message to contextualize spurious prior error() message.
Slurm checks for system variable 'wsrep_on' in mysql, which was only added
as of MariaDB versions >= 10.1 and some builds of MySQL.  Checking for
it on systems that don't have it defined results in a spurious error
"1193 Unknown system variable 'wsrep_on'".  Adding some context here to
assure users that this error is harmless and the feature will simply be
disabled.

Bug 19385