Prevent rem_node from always being set to 1 in eval_nodes_select_cores.
gres_filter_sock_core needs rem_node to be the number or nodes remaining to
be added to the allocation including itself. This was not an issue before
because nothing was using the value, but now gres_filter_sock_core() uses
it to calculate max_gres when using gres_per_job.

Bug 14227