X11 - Accept magic cookies without display number
While using xauth list $DISPLAY it will also show the cookies that do
not have associated a display number, as these are cookies that can be
used with any display number, we are not currently accepting this
cookies as in the regex cookie_pattern, we have the modifier '+' in the
display number part, indicating that 1 or more need to match. This
commit changes the '+' to '*' to accept cookies with the default
display.

Bug 19543