switch/hpe_slingshot - Drain nodes on failed CXI service delete.
Change the response to a failed switch_g_job_postfini() call so that the
node is drained. This helps when the Slingshot driver can't delete a CXI
service.  The switch/hpe_slingshot plugin tries 5 times to delete the
service, then gives up. While an epilog can be set up to loop on service
deletion, that epilog only works on exclusive partitions. Handling it here
will work for shared partitions as well.

Ticket 19601.