cgroup/v2 - Require dbus-1 version >= 1.11.16.
The cgroup/v2 plugin requires a version of dbus-1 >= 1.11.16 so that all
used symbols are available. This commit modifies the configure script so
that dbus-1 presence isn't just checked but also the required version is
satisfied. Otherwise, the plugin won't be built.

Some distributions like Amazon Linux 2 OS are shipped with:
- kernel-headers 5.10 (cgroup/v2 eBPF functionality requires >= 5.7)
- dbus-1 1.10.24 (cgroup/v2 dbus functionality requires >= 1.11.16)

So before this commit the configure script on such an example distribution
wouldn't emit any warning, but at build time the compiler would complain due
to missing symbols.

Bug 19615