Docs - Add versions in the cgroup_v2 requirements
Added these minimum versions as these are needed to properly build
cgroup_v2 plugin.

kernel header needs to be at least 5.7 as the newest define required for
bpf is contained in linux kernel commit 1aae4bdd78799, which ships with
5.7

dbus needs to be on 1.11.16 as the needed symbols are contained in that
version, see https://bugs.freedesktop.org/show_bug.cgi?id=101568 and
the dbus NEWS entry for 1.11.16 Enhancements entry.

Bug 19615