Make gres_filter_sock_core not reducing avail_core bits in most cases.
Unless it calculates that the available cores on a socket are 0 don't
actually reduce the avail_core bitmap. The number of available cores on
each socket will be saved in
topo_eval->avail_res_arrayu[node_i]->avail_cores_per_sock.
This will be used in the following commit to reduce the bitmap after the
nodes are selected. This is to allow for the removal of restricted cores if
it is later found during the node selection process that it needs less gres
then currently though.

Bug 14227