Don't set job's priority in main scheduler when the job is running
This leaves the priority on the job set from the partition that started
the job. This is seen with squeue -o priority after the job has started.

e.g.
Priorities:
p1 1
p2 2
p3 3

sbatch -p p1,p2,p3

1. start job on p3
2. job's priority is set to 2 from the p2 job_queue_rec. The scheduled
2a. the scheduler continues because the job isn't pending.
3. job's priority is set to 1 from the p1 job_queue_rec. The scheduled
3a. the scheduler continues because the job isn't pending.

The commit delays setting the job's priority after the check of whether
the job is running or not.

Bug 17517