Fix signaling job arrays with slurm_kill_jobs
This fixes a few problems with job arrays.
The square bracket job array expression was not parsed and resulted in an
error. For example:

$ scancel -v 710_[1,3-4] --ctld
scancel: error: Kill job error on job id 710_1: Invalid job id specified

The previous commits adds support for job array expressions in
the functions foreach_job_by_id[_ro]. Pending jobs that are not split from
the job array are handled in an identical way to job_str_signal().
Jobs that have been split from the job array are signaled in the usual way.

In addition to fixing job array expressions, this commit fixes signaling
pending job array tasks that are not split from the meta array record.
Previously, attempting to signal such a task would result in signaling the
whole job array. This commit handles such a task in the same way as a job
array expression.

Bug 18464

Signed-off-by: Nathan Rini <nate@schedmd.com>