Change extern slurm_persist_conn_open_without_init() function to static
With the previous commits, slurm_persist_conn_open_without_init() is no
longer used anywhere, and should not be used anywhere.

All persistent connections need to be started with
REQUEST_PERSIST_INIT[_TLS] in order to determine whether TLS will be used in
the connection or not.

slurm_persist_conn_open_without_init() is renamed to _open_persist_conn().

Bug 19121