Initialize the slurmdbd backup persistent connection.
Needed to bring up the tls connection to ensure communication
is working as expected.

Set r_uid to slurm_user_id - this does assume the slurmdbd is operating
as SlurmUser on both primary and backup, but that is the only supported
configuration and should be a reasonably safe assumption here.

Bug 19121