data_t - Auto convert in data_check_match()
Avoid data_check_match() failing when data can be converted to the same
type and matches. This explicitly excludes conversion to DATA_TYPE_NULL as
that conversion just drops the value and will always match.

Use data_get_type() to match the types. This ensures that inline strings and
xmalloc strings match.

Bug 19225